+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [unreleased]
+- Introduced `newDeriveKeyHasher(Hasher => Unit)`.
 - Introduced `update(Short)`, `update(Int)` and `update(Long)`.
 
 ## [2.2.0] - 2021-01-05